# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_eleven-am_voice-backend_internal_dto.MeResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.png
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: usr_abc123
        type: string
      is_developer:
        example: false
        type: boolean
      name:
        example: John Doe
        type: string
    type: object
...
  github_com_eleven-am_voice-backend_internal_shared.APIError:
    properties:
      code:
        example: invalid_request
        type: string
      details:
        type: object
      message:
        example: Invalid request body
        type: string
    type: object
host: api.voice.example.com
info:
  contact: {}
  description: API server for voice agent platform
  title: Voice Backend API
  version: 1.0.0
paths:
...
  /auth/github:
    get:
      description: Redirects the user to GitHub OAuth consent page
      parameters:
      - description: URL to redirect after successful login
        in: query
        name: redirect_uri
        type: string
      responses:
        "307":
          description: Redirect to GitHub OAuth
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Initiate GitHub OAuth login
      tags:
      - auth
  /auth/github/callback:
    get:
      description: Handles the callback from GitHub OAuth and creates user session
      parameters:
      - description: OAuth state parameter
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      responses:
        "307":
          description: Redirect to redirect_uri or /
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: GitHub OAuth callback
      tags:
      - auth
  /auth/google:
    get:
      description: Redirects the user to Google OAuth consent page
      parameters:
      - description: URL to redirect after successful login
        in: query
        name: redirect_uri
        type: string
      responses:
        "307":
          description: Redirect to Google OAuth
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Initiate Google OAuth login
      tags:
      - auth
  /auth/google/callback:
    get:
      description: Handles the callback from Google OAuth and creates user session
      parameters:
      - description: OAuth state parameter
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      responses:
        "307":
          description: Redirect to redirect_uri or /
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Google OAuth callback
      tags:
      - auth
  /auth/logout:
    post:
      description: Clears the user session and logs out
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the currently authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Get current user
      tags:
      - auth
  /auth/me/developer:
    post:
      description: Upgrades the current user to developer status
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Become a developer
      tags:
      - auth
...
securityDefinitions:
  APIKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  SessionAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"